Adding two routes file
